feat: support multiple warframes/weapons being given in the same request
